'create the enumeration for the variable below (GameState
currentState)' updated.

'This JPanel fires a MouseEvent on MouseClicked so add required event
listener to 'this' updated.

'Create a new instance of the game "Board"class. HINT check the
variables above for the correct name' updated.

'Call the method to initialise the game board' updated.

'Create the new GameMain panel and add it to the frame' updated.

'Set the default close operation of the frame to exit_on_close' updated.

'Use the status bar to display the message "X"'s Turn' updated.

' Use the status bar to display the message "O"'s Turn' updated.

'Check which player has won and update the currentstate to the
appropriate gamestate for the winner' updated.

'set the currentstate to the draw gamestate' updated.

'redraw the graphics on the UI' updated.